+++ ui/src/components/Sidebar/MobileSidebar.tsx
@@ -6,7 +6,7 @@ import useGroupSort from '@/logic/useGroupSort';
 import { useNotifications } from '@/notifications/useNotifications';
 import Notifications from '@/notifications/Notifications';
 import GroupNotification from '@/notifications/GroupNotification';
-import useNavStore from '../Nav/useNavStore';
+import useNavStore from '../../nav/useNavStore';
 import SidebarSorter from './SidebarSorter';
 import NavTab from '../NavTab';
 import GroupIcon from '../icons/GroupIcon';
