+++ ui/src/groups/GroupAdmin/GroupAdmin.tsx
@@ -2,7 +2,7 @@ import cn from 'classnames';
 import React from 'react';
 import { NavLink, Outlet } from 'react-router-dom';
 import CaretLeftIcon from '../../components/icons/CaretLeftIcon';
-import useNavStore from '../../components/Nav/useNavStore';
+import useNavStore from '../../nav/useNavStore';
 import { useIsMobile } from '../../logic/useMedia';
 import { useAmAdmin, useRouteGroup } from '../../state/groups/groups';
