+++ ui/src/groups/GroupAdmin/GroupInfoEditor.tsx
@@ -14,7 +14,7 @@ import {
   ViewProps,
 } from '@/types/groups';
 import { useNavigate } from 'react-router';
-import useNavStore from '@/components/Nav/useNavStore';
+import useNavStore from '@/nav/useNavStore';
 import { getGroupPrivacy } from '@/logic/utils';
 import GroupInfoFields from '../GroupInfoFields';
 import PrivacySelector from '../PrivacySelector';
