--- ui/src/pages/Search.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Search() {
-  return <div className="w-full" />;
-}
